refactor: extract methods to CardTemplate
These belonged in the class, not in HtmlGenerator
