+++ AnkiDroid/src/main/java/com/ichi2/anki/cardviewer/HtmlGenerator.kt
@@ -26,11 +26,6 @@ import com.ichi2.libanki.Card
 import com.ichi2.libanki.Collection
 import com.ichi2.libanki.Sound
 import com.ichi2.libanki.stripAvRefs
-import timber.log.Timber
-import java.io.BufferedReader
-import java.io.IOException
-import java.io.InputStream
-import java.io.InputStreamReader
 
 class HtmlGenerator(
     private val typeAnswer: TypeAnswer,
@@ -68,7 +63,7 @@ class HtmlGenerator(
         ): HtmlGenerator {
             val preferences = context.sharedPrefs()
             val cardAppearance = CardAppearance.create(ReviewerCustomFonts(), preferences)
-            val cardHtmlTemplate = loadCardTemplate(context)
+            val cardHtmlTemplate = CardTemplate.load(context)
             val showAudioPlayButtons = !col.config.getBool(ConfigKey.Bool.HIDE_AUDIO_PLAY_BUTTONS)
             return HtmlGenerator(
                 typeAnswer,
@@ -77,42 +72,6 @@ class HtmlGenerator(
                 showAudioPlayButtons,
                 context.resources
             )
-        }
-
-        /**
-         * Load the template for the card
-         */
-        private fun loadCardTemplate(viewer: Context): CardTemplate {
-            try {
-                val data = convertStreamToString(viewer.assets.open("card_template.html"))
-                return CardTemplate(data)
-            } catch (e: IOException) {
-                Timber.w(e)
-                throw RuntimeException(e)
-            }
-        }
-
-        /**
-         * Converts an InputStream to a String.
-         *
-         * @param input InputStream to convert
-         * @return String version of the InputStream
-         */
-        private fun convertStreamToString(input: InputStream?): String {
-            var contentOfMyInputStream = ""
-            try {
-                val rd = BufferedReader(InputStreamReader(input), 4096)
-                var line: String?
-                val sb = StringBuilder()
-                while (rd.readLine().also { line = it } != null) {
-                    sb.append(line)
-                }
-                rd.close()
-                contentOfMyInputStream = sb.toString()
-            } catch (e: Exception) {
-                Timber.w(e)
-            }
-            return contentOfMyInputStream
         }
     }
 }
