+++ AnkiDroid/src/test/java/com/ichi2/anki/cardviewer/CardTemplateTest.kt
@@ -30,7 +30,7 @@ class CardTemplateTest {
         val style = "bar"
         val cardClass = "baz"
         val script = "script"
-        val result = CardTemplate(data).render(content, style, script, cardClass)
+        val result = CardTemplate(data).render(content, style, script, cardClass).html
         assertThat(result, equalTo(data.replace("::content::", content).replace("::style::", style).replace("::class::", cardClass).replace("::script::", script)))
     }
